Performance improvement in data binding transaction history
dgMain.ScrollViewer.CanContentScroll was set to false, which apparently delegates the job of scrolling to the ScrollViewer itself, bypassing the more intelligent optimizations a DataGrid can make. Setting to true substantially reduced the noticable lag in searching, though it may lead to odd rendering if a single row is very tall relative to the entire pane.